feat(editor): Redact secrets and PII from AI Assistant frontend telemetry (no-changelog)
Frontend events reach RudderStack *and* PostHog straight from the browser,
so no backend hop sees them and redaction has to happen at the call site.

Extracts the redaction engine into @n8n/utils so the browser can run the
same policy as the backend, and moves the telemetry egress policy into
@n8n/telemetry, shared by both.

Applies it to the AI Assistant properties that carry free text: the user's
typed feedback, their typed answers to agent question cards, the agent's own
description of an action it wants to take, and the host it asks to reach.
Identifier keys are exempted, so the dashboard join keys are untouched.

Also makes redactDeep withhold an over-depth subtree regardless of
redactSensitiveKeys — the guard previously passed the subtree through
unscanned in the non-key mode.
