+++ packages/frontend/editor-ui/src/features/ai/instanceAi/__tests__/InstanceAiConfirmationPanel.test.ts
@@ -559,6 +559,56 @@ describe('InstanceAiConfirmationPanel telemetry', () => {
 			);
 		});
 
+		it('scrubs secrets and PII from the typed answer and the question', async () => {
+			injectPendingConfirmation(thread, {
+				requestId: 'req-text-pii',
+				severity: 'info',
+				message: 'Which address should the invoice go to, jane.doe@example.com?',
+				inputType: 'text',
+			});
+			vi.spyOn(thread, 'confirmAction').mockResolvedValue(true);
+
+			const { container } = renderComponent({ props: { kind: 'inline' } });
+			const input = container.querySelector('input[type="text"]') as HTMLInputElement;
+			await userEvent.type(input, 'use billing@acme.com');
+			await userEvent.keyboard('{Enter}');
+
+			expect(mockTelemetryTrack).toHaveBeenCalledWith(
+				'User finished providing input',
+				expect.objectContaining({
+					provided_inputs: [
+						{
+							label: 'Which address should the invoice go to, [REDACTED]?',
+							question: 'Which address should the invoice go to, [REDACTED]?',
+							input_type: 'text',
+							options: [],
+							option_chosen: 'use [REDACTED]',
+						},
+					],
+				}),
+			);
+		});
+
+		it('keeps the thread and instance identifiers the dashboards join on', async () => {
+			injectPendingConfirmation(thread, {
+				requestId: 'req-text-ids',
+				severity: 'info',
+				message: 'What name for the workflow?',
+				inputType: 'text',
+			});
+			vi.spyOn(thread, 'confirmAction').mockResolvedValue(true);
+
+			const { container } = renderComponent({ props: { kind: 'inline' } });
+			const input = container.querySelector('input[type="text"]') as HTMLInputElement;
+			await userEvent.type(input, 'My Workflow');
+			await userEvent.keyboard('{Enter}');
+
+			expect(mockTelemetryTrack).toHaveBeenCalledWith(
+				'User finished providing input',
+				expect.objectContaining({ thread_id: 'thread-1', instance_id: 'test-instance-id' }),
+			);
+		});
+
 		it('tracks text skip with input_type and question', async () => {
 			injectPendingConfirmation(thread, {
 				requestId: 'req-text-skip',
